+++ src/blocks/godslayerakp/discord/roles/index.js
@@ -1,3 +1,4 @@
 import "./create role then"
 import "./create_role_then_real"
 import "./created_role"
+import "./get roles list"
