+++ src/toolbox.js
@@ -2655,6 +2655,7 @@ export default (Blockly, value, searching) => {
             <block type="s4d_get_all_role_role"></block>
             <block type="jg_roles_get_all_member_roles_then_for_each_do"/>
             <block type="jg_roles_get_all_member_roles_then_for_each_do_role"/>
+            <block type="gsa_get_all_roles_on_member"></block>
             <block type="hig_role"/>
             <!--
             <block type="monaco_members_with_role"/>
